Voici le dernier pull version complet ONT + metaspades

+++ sources/assembly/metaflye_wraper.sh
@@ -4,11 +4,21 @@
 # Get parameters
 run="$1"
 output_directory="$2"
-
+technology="$3"
 
 mkdir -p "$output_directory"/tmp
-flye --meta -t $(nproc) --out-dir "$output_directory"/tmp/ --pacbio-hifi "$run"
+
+# Select read type depending on sequencing technology
+if [ "$technology" = "ont" ]; then
+    READ_TYPE="--nano-hq"
+else
+    READ_TYPE="--pacbio-hifi"
+fi
+
+flye --meta -t $(nproc) --out-dir "$output_directory"/tmp/ $READ_TYPE "$run"
 
 mv "$output_directory"/tmp/assembly.fasta "$output_directory"
-rm -rf "$output_directory"/tmp/
+if [ -f "$output_directory/assembly.fasta" ]; then
+    rm -rf "$output_directory"/tmp/
+fi
